+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add the Reduce function on a RangeCollection<T>, returns a new RangeCollection<T> where all overlapping and touching Ranges have been merged and empty Ranges have been removed
+
 ## [0.2.0] - 2019-01-02
 ### Added
 - The Exclusive method, that returns a Tuple of two Ranges that represent the parts they do not have in common, including the ^ operator